Rework the font cache to better handle fallbacks
The previous version of the font cache would always fail lookups in the
font cache when a fallback font was used, leading to having to re-create
the SkTypeface every frame.

This reworks the font cache to store the typefaces in the font entries,
which means that the separate font cache map can be dropped. As a
result, it's now possible to store the results of fallback lookups in
the cache.

Although the cache entries must now be guarded by a mutex, the
separate SkTypeface lookup map from before also had to be, so the
amount of delays due to locking should be largely equivalent.