Handle extreme radii in add_round_rect
This originated as a follow on from a fix to cycfi/elements. We wish to
ensure that the radius for a round rectangle does not exceed half of the
minimum dimension of the rectangle. I have chosen to introduce the
method `add_round_rect_impl` to both `canvas` and `path` so that
regardless of the backend, we get consistent behaviour when the radius
is too large.

Note: I believe the implementation in skia would have given this result
already, but to me it seems preferable to control this instead of
allowing implementation-defined behaviour.